+++ src/bindings/gpucomputepipeline.h
@@ -16,7 +16,7 @@ class GPUComputePipeline : public interop::GPUComputePipeline {
   operator wgpu::ComputePipeline() const { return pipeline_; }
 
   interop::Interface<interop::GPUBindGroupLayout> getBindGroupLayout(
-      Napi::Env, unsigned long index) override;
+      Napi::Env, uint32_t index) override;
   std::optional<std::string> getLabel(Napi::Env) override;
   void setLabel(Napi::Env, std::optional<std::string> value) override;
